requirements.txt
transformers
torch
Flask
cerebras-cloud-sdk